Function to download CHIRPS raster data global on daily and monthly scales.

Function to download CHIRPS raster data global on daily and monthly scales.

download_CHIRPS(
  tres = NULL,
  sres = NULL,
  years = 1981,
  dsn = getwd(),
  format = NULL,
  cores = parallel::detectCores() - 1,
  verbose = FALSE,
  roi = c(-86, -66, -20, 2)
)

Arguments

tres

character. Temporal resolution. "daily" or "monthly"

sres

numeric. Spatial resolution. 0.05 or 0.25. Note that monthly data is only available at a resolution of 0.05 degrees.

years

numeric. Years Date(s) for which you want to download CHIRPS data.

dsn

path. Directory to which you want to store the downloaded data.

format

character. Download format of the files. "tif" or "netcdf" are only available.

cores

numeric. Number of cores to be used for the download. By default, it automatically detects how many cores there are on your computer and subtracts 1 to prevent it from slowing down.

verbose

logical. If TRUE this will show the task process for the parallel process.

roi

numeric. Vector with xmax, xmin, ymax, ymin values for a region for subsetting the data.

References

https://data.chc.ucsb.edu/products/CHIRPS-2.0/

See also

Author

Crhistian Cornejo

Examples

if (FALSE) {
# roi = c(-86,-66,-20,2)
# download_CHIRPS(tres = 'daily',sres = 0.25,format = 'tif',years = 2000,roi = roi)
}