Download available data for PERÚ at 0.1 x 0.1 resolution of grid and subsetting to a roi area if require.

Download available data for PERÚ at 0.1 x 0.1 resolution of grid and subsetting to a roi area if require.

download_RAIN4PE(
  dsn = getwd(),
  openPDF = TRUE,
  shp = NULL,
  name_roi = "Cuenca_Mayo"
)

Arguments

dsn

path. Directory to which you want to store the downloaded data

openPDF

logical. If TRUE the PDF of the reference paper will be open. FALSE just donwload the grid data.

name_roi

character. Name of the area or basin with which the subsetting to the grid will be done.

References

Fernandez-Palomino, Carlos Antonio; Hattermann, Fred F.; Krysanova, Valentina; Lobanova, Anastasia; Vega-Jácome, Fiorella; Lavado, Waldo; Santini, William; Aybar, Cesar; Bronstert, Axel (2021): Rain for Peru and Ecuador (RAIN4PE). V. 1.0. GFZ Data Services. https://doi.org/10.5880/pik.2020.010

See also

Author

Crhistian Cornejo

Examples

if (FALSE) {
## donwloading data and subsetting for RAIN4PE
shp <- 'G:/Mi unidad/MyPackages-R/getSPPs/test/SHPs/cuenca_mayo.shp'
download_RAIN4PE(openPDF = FALSE, shp = shp, name_roi = 'Cuenca_Test')
}