Download 14 climatic variables simultaneously from the TERRACLIMATE database at 4km resolution.

Download 14 climatic variables simultaneously from the TERRACLIMATE database at 4km resolution.

download_TERRACLIMATE(
  vars_id = c("aet", "def", "swe", "q", "soil", "PDSI", "pet", "ppt", "srad", "tmax",
    "tmin", "vap", "vpd", "ws"),
  years = 1958:2021,
  dsn = getwd(),
  cores = parallel::detectCores() - 1
)

Arguments

vars_id

character. Name of variables what you want to download. for DEFAULT it download all variables available.

years

integer. Years from which you want to download the information, by DEFAULT it downloads all the available years of TERRACLIMATE of all the variables.

dsn

character. Path where the output folders and files will be saved.

cores

numeric. Necessary parameter to download multiple variables in parallel.

Value

Donwload TERRACLIMATE data sets to output directory.

References

Abatzoglou, J.T., S.Z. Dobrowski, S.A. Parks, K.C. Hegewisch, 2018, Terraclimate, a high-resolution global dataset of monthly climate and climatic water balance from 1958-2015, Scientific Data,

See also

Examples

if (FALSE) {
if(interactive()){
download_TERRACLIMATE(vars_id  = c('aet','ppt'), years = 2000:2002, cores = 4)
 }
}