Function to get codes of senamhi water stations.

Function to get codes of senamhi water stations.

get_codes(
  stations_names = NULL,
  type = c("Met", "Hid"),
  get_LongLat = TRUE,
  plotStations = TRUE
)

Arguments

stations_names

character. Names of the stations we want to download and get their code.

type

character. Type of stations from which the codes will be returned.

get_LongLat

logical. If TRUE Longitude and Latitude will added to the output dataframe. This needed for plotStations function.

plotStations

logical. TRUE will plot a graph of ubications of the stations and his altitude.

Value

If get_LongLat and plotStations are TRUE a list with the table and a plot of the stations is returned, if both are FALSE, only a two-column dataframe will be returned.

References

this functions make usefull easy wrangling download data from the package senamhiR from source("https://gitlab.com/snippets/1793256/raw");install("senamhiR")

See also

Author

Crhistian Cornejo

Examples

 if (FALSE) {
#example for get the codes and pre-process for get_SENAMHI_data function if return is a list
names <- c('Asuncion','Augusto','llica','honda','llagaden',
           'namora','jesus','la victoria','celendin')
codes <- get_codes(stations_names = names,
                  type = 'Met',
                  get_LongLat = TRUE,
                  plotStations = TRUE)

 codes_for_function <- do.call(rbind, lapply(codes[1], data.frame, stringsAsFactors=FALSE)) %>%
 dplyr::arrange(StationID) %>%
 dplyr::select(StationID) %>%
 purrr::as_vector()


}