Function extracts grid data from climate raster bricks and make a database table for later import in RSMinerve platform.

Function extracts grid data from climate raster bricks and make a database table for later import in RSMinerve platform.

grid2RSMinerve(
  grid,
  subbasins,
  var_type = "Precipitation",
  crs = 32717,
  DEM,
  dsn = getwd(),
  dates_rs = list(),
  interpolation_method = "Linear",
  cores = parallel::detectCores() - 1
)

Arguments

grid

raster brick or raster stack

subbasins

PARAM_DESCRIPTION

var_type

Either 'Temperature' or 'Precipitation'.Default: 'Precipitation'

crs

4 digit crs code to ensure projection consistency between raster and shapefile. Default: 32717

DEM

Digitial Elevation Model (raster) for extracting z values.

dsn

path directory to which you want to store the downloaded data.

dates_rs

list with star_date and end_date of where the raster brick have information, Default: list()

interpolation_method

type of interpolation reading for RSMinerve, Default: 'Linear'

cores

numeric. Number of cores to be used for the download. By default, it automatically detects how many cores there are on your computer and subtracts 1 to prevent it from slowing down.

See also

Examples

if (FALSE) {
if(interactive()){
DEM <- raster::raster('usefull_data/pruebas/DEM_alos_JT.tif')
subbasins <- sf::st_read('usefull_data/pruebas/Subcuencas_v006.shp')
grid_pr <- raster::brick('usefull_data/pruebas/prec_diaria.nc')

grid2RSMinerve(grid = grid_pr,
               subbasins = subbasins,
               var_type = 'Precipitation',
               crs = 32717,DEM = DEM,
               dsn = getwd(),
               interpolation_method = 'Linear',
               dates_rs = list(Start = '1981-01-01', end = '2016-12-31'))
 }
}