Function to subset CMORPH raster data to a roi extension.

Function to subset CMORPH raster data to a roi extension.

subsetting_CMORPH(
  path_files = getwd(),
  roi = c(-86, -66, -20, 2),
  vars = list(Lat = "lat", Long = "lon", main_var = "cmorph", time = "time"),
  file_out = sub("[.]nc", "_subsetting.nc", path_files),
  cores = parallel::detectCores() - 1
)

Arguments

path_files

list. Path to the files where the CMORPH.nc are downloaded.

roi

numeric. Vector with xmax, xmin, ymax, ymin values for a region for subsetting the data.1

vars

list. Variables necessary to perform the subsetting, do not change it for CMORPH grids.

file_out

path. Path where the new files will be saved. By default they are saved in the same path as the downloaded files and are automatically renamed.

cores

numeric. Number of cores to be used for the download. By default, it automatically detects how many cores there are on your computer and subtracts 1 to prevent it from slowing down.

Value

subsetting tif.

References

Joyce, R. J., J. E. Janowiak, P. A. Arkin, and P. Xie, 2004: CMORPH: A method that produces global precipitation estimates from passive microwave and infrared data at high spatial and temporal resolution.. J. Hydromet., 5, 487-503.

Author

Crhistian Cornejo

Examples