Get information from multiple station and variables from Servicio Nacional de Meteorología e Hidrológia del Perú (SENAMHI).

Get information from multiple station and variables from Servicio Nacional de Meteorología e Hidrológia del Perú (SENAMHI).

get_SENAMHI_data(
  cod = NULL,
  start_date = NULL,
  end_date = NULL,
  convert2monthly = FALSE,
  convert2daily = FALSE,
  na.rm = T,
  vars = c("pr", "tmp", "wind")
)

Arguments

cod

character. Codes of the station you want to download, to see the codes use link{get_codes}. For the moment just for Conventional stations data are available.

start_date

character. Start date to filter the data what you want to download. Please write as 'yyyy-mm-dd'.

end_date

character. End date to filter the data what you want to download. Please write as 'yyyy-mm-dd'.

convert2monthly

logical.If TRUE will return the data in monthly values.

convert2daily

logical. If TRUE will return the data in daily values (this just works for precipitation data).

na.rm

logical. TRUE for remove NA's values from data if convert2monthly is TRUE. FALSE will return the monthly mean values convert2monthly NA's.

vars

character. Variables to found in the database from SEAMHI and return.

Value

A matrix with each station data of variable request, values of precipitation on (mm/hour, mm/day), temperature values on (°C) and wind data on direction and velocity unit (m/s)

References

this functions make usefull easy wrangling download data from the package senamhiR from source("https://gitlab.com/snippets/1793256/raw");install("senamhiR")

See also

Author

Crhistian Cornejo

Examples

if (FALSE) {
#example for get the codes and pre-process for get_SENAMHI_data function
names <- c('Asuncion','Augusto','llica',
'honda','llagaden','namora',
'jesus','la victoria','celendin')
codes <- get_codes(stations_names = names,
                  type = 'Meteorological',
                  get_LongLat = TRUE,
                  plotStations = TRUE)

 codes_for_function <- do.call(rbind, lapply(codes[1], data.frame,
 stringsAsFactors=FALSE)) %>%
 dplyr::arrange(StationID) %>%
 dplyr::select(StationID) %>%
 purrr::as_vector()
#for temperature
Data_T <- get_SENAMHI_data(cod = codes_for_function,
                          start_date = '2000-01-01',
                          end_date = '2015-12-31',
                          convert2monthly = FALSE,
                          convert2daily = FALSE,
                          na.rm = FALSE,
                          vars = 'tmp')
#getting the names matching names as vector for replace on the export
real_names <- do.call(rbind, lapply(codes[1], data.frame, stringsAsFactors=FALSE)) %>%
 dplyr::filter(StationID %in%
 codes_for_function) %>%
 dplyr::arrange(StationID) %>%
 dplyr::select(1) %>% as.data.frame() %>% purrr::as_vector(.)

nombres_est <- as.vector(real_names)
colnames(Data_T) <- c('Date',rep(paste0('tmax_',nombres_est),1),
rep(paste0('tmin_',nombres_est),1))
#for precipitation
Data_P <- get_SENAMHI_data(cod = codes_for_function,
                          start_date = '2000-01-01',
                          end_date = '2015-12-31',
                          convert2monthly = FALSE,
                          convert2daily = FALSE,
                          na.rm = FALSE,
                          vars = 'pr')
#getting the names matching names as vector for replace on the export of precipitation dataframe
real_names_p <- do.call(rbind, lapply(codes[1], data.frame, stringsAsFactors=FALSE)) %>%
 dplyr::filter(StationID %in% codes_for_function) %>%
 dplyr::arrange(StationID) %>% dplyr::select(1) %>% as.data.frame() %>% purrr::as_vector(.)

nombres_est_p <- as.vector(real_names_p)
colnames(Data_P) <- c('Date',rep(paste0('Prec07_',nombres_est),1),
                             rep(paste0('Prec19_',nombres_est),1))
write.csv(Data_P, '../SENAMHI_Prec_2000_2015_subdaily.csv')
}